Don't count a pod as ready until it joins the quorum
By default the readiness check in a pod will assume once the daemon is
running it is ready. However consul can take some amount of time to
remove the old node and add a new one, and in that time if the
StatefulSet has killed another pod, then our cluster can get wedged
without a quorum.

By allowing unready pods to show up in the DNS and treating pods as
unready until they have joined quorum, we ensure that only one pod is
removed from the cluster at a time.