fix: move client geolocation earlier to avoid shield
With the current implementation, any requests coming through a Shield
node will be annotated with the Shield's geo information instead of the
originating client's geo data.

By moving this earlier, to the one place we interact with the `client`
object prior to further processing, we ensure that we set the right data
at the right time.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>